explicit cast on ros time now

+++ mav_control_interface/src/state_machine.h
@@ -574,7 +574,7 @@ private:
     template<class FSM, class SourceState, class TargetState>
     bool operator()(const OdometryWatchdog& evt, FSM& fsm, SourceState&, TargetState&)
     {
-      return (std::abs(ros::Time::now().toNSec() - fsm.current_state_.timestamp_ns)) > kOdometryOutdated_ns;
+      return std::abs(static_cast<int64_t>(ros::Time::now().toNSec()) - fsm.current_state_.timestamp_ns) > kOdometryOutdated_ns;
     }
   };
